convert jsdoc template to snippet
